bronto: remove generic `track` method, for #361
As it stands now, bronto is only being used for tracking "completed order". So the integration is only going to support that for now. If we run into a new use case for bronto, we can go from there, but it makes more sense just to implement only what's necessary atm (if the integration is a little more difficult to figure out).